chore: move to its own file and update test
