+++ packages/nuxt/src/app/composables/hydrate.ts
@@ -2,36 +2,6 @@ import { useNuxtApp } from '../nuxt'
 import type { NuxtPayload } from '../nuxt'
 
 /**
- * An SSR-friendly utility to call a method once
- * @param key a unique key ensuring the function can be properly de-duplicated across requests
- * @param fn a function to call
- * @see https://nuxt.com/docs/api/utils/once
- */
-export function once (key?: string, fn?: (() => any | Promise<any>)): Promise<void>
-export function once (fn?: (() => any | Promise<any>)): Promise<void>
-export async function once (...args: any): Promise<void> {
-  const autoKey = typeof args[args.length - 1] === 'string' ? args.pop() : undefined
-  if (typeof args[0] !== 'string') { args.unshift(autoKey) }
-  const [_key, fn] = args as [string, (() => any | Promise<any>)]
-  if (!_key || typeof _key !== 'string') {
-    throw new TypeError('[nuxt] [once] key must be a string: ' + _key)
-  }
-  if (fn !== undefined && typeof fn !== 'function') {
-    throw new Error('[nuxt] [once] fn must be a function: ' + fn)
-  }
-  const nuxt = useNuxtApp()
-  if (import.meta.server) {
-    await fn()
-    nuxt.payload.once[_key] = true
-    return
-  }
-  if (nuxt.isHydrating && nuxt.payload.once[_key]) {
-    return
-  }
-  await fn()
-}
-
-/**
  * Allows full control of the hydration cycle to set and receive data from the server.
  * @param key a unique key to identify the data in the Nuxt payload
  * @param get a function that returns the value to set the initial data
