+++ packages/nuxt/src/app/composables/index.ts
@@ -15,7 +15,8 @@ export {
 export { defineNuxtComponent } from './component'
 export { useAsyncData, useLazyAsyncData, useNuxtData, refreshNuxtData, clearNuxtData } from './asyncData'
 export type { AsyncDataOptions, AsyncData } from './asyncData'
-export { once, useHydration } from './hydrate'
+export { useHydration } from './hydrate'
+export { once } from './once'
 export { useState, clearNuxtState } from './state'
 export { clearError, createError, isNuxtError, showError, useError } from './error'
 export type { NuxtError } from './error'
